Updated user id to Long and bigint.

+++ src/main/java/ua/lastbite/userservice/exception/user/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package ua.lastbite.userservice.exception.user;
 
 public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(Integer id) {
+    public UserNotFoundException(Long id) {
         super("User with ID " + id + " not found");
     }
 }
